+++ source/Private/Get-SEPCloudAPIData.ps1
@@ -1,12 +1,10 @@
-function Get-SEPCloudAPIData
-{
+function Get-SEPCloudAPIData {
     [CmdletBinding()]
     param (
         $endpoint
     )
 
-    process
-    {
+    process {
         $api = @{
             'Example'                                  = @{
                 '1.0' = @{
@@ -512,13 +510,62 @@ function Get-SEPCloudAPIData
                     ObjectTName = 'SEPCloud.DeviceCommandResponse'
                 }
             }
+            'Update-AllowListPolicyByFileHash'         = @{
+                '1.0' = @{
+                    Description = 'perform partial update of Allow list policy by filename'
+                    URI         = '/v1/policies/allow-list/{id}/versions/{id}'
+                    Method      = 'Patch'
+                    Body        = @{
+                        add = @{
+                            applications = @(
+                                @{
+                                    processfile = @{
+                                        sha2 = 'sha2'
+                                        name = 'name'
+                                    }
+                                }
+                            )
+                        }
+                    }
+                    Query       = ''
+                    Result      = ''
+                    Success     = ''
+                    Function    = 'Update-AllowListPolicyByFileHash'
+                    ObjectTName = 'SEPCloud.update-policy-response'
+                }
+            }
+            'Update-AllowListPolicyByFileName'         = @{
+                '1.0' = @{
+                    Description = 'perform partial update of Allow list policy by filename'
+                    URI         = '/v1/policies/allow-list/{id}/versions/{id}'
+                    Method      = 'Patch'
+                    Body        = @{
+                        add = @{
+                            windows = @{
+                                files = @(
+                                    @{
+                                        pathvariable = 'pathvariable'
+                                        path         = 'path'
+                                        scheduled    = 'scheduled'
+                                        features     = 'features'
+                                    }
+                                )
+                            }
+                        }
+                    }
+                    Query       = ''
+                    Result      = ''
+                    Success     = ''
+                    Function    = 'Update-AllowListPolicyByFileName'
+                    ObjectTName = 'SEPCloud.update-policy-response'
+                }
+            }
         }
 
         # Use the latest version of the API endpoint
         $version = $api.$endpoint.Keys | Sort-Object | Select-Object -Last 1
 
-        if ($null -eq $version)
-        {
+        if ($null -eq $version) {
             $ErrorSplat = @{
                 Message      = "No matching endpoint found for $Endpoint that corresponds to the current cluster version."
                 ErrorAction  = 'Stop'
@@ -526,9 +573,7 @@ function Get-SEPCloudAPIData
                 Category     = 'ObjectNotFound'
             }
             Write-Error @ErrorSplat
-        }
-        else
-        {
+        } else {
             Write-Verbose -Message "Selected $version API Data for $endpoint"
             return $api.$endpoint.$version
         }